requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
pydantic-settings==2.4.0
python-multipart==0.0.22
alibabacloud_dingtalk==2.0.75
alibabacloud_tea_openapi==0.3.8
redis==5.0.4
SQLAlchemy==2.0.30
